Refactor App and context
